requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.38.0
huggingface-hub==0.10.1
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.1
notebook_shim==0.2.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
pandas-stubs==1.2.0.62
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.13.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2022.9.13
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentence-transformers==2.0.0
sentencepiece==0.1.97
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.3.2.post1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.10.3
torch==1.13.0
torchaudio==0.13.0
torchtext==0.14.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.15.0
typing_extensions==4.4.0
urllib3 @ file:///croot/urllib3_1666298941550/work
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
zipp==3.9.0
